docs/guides/Haven-lore/keeper-bot/requirements.txt
discord.py>=2.3.0
aiofiles>=23.2.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pillow>=10.0.0
asyncio-throttle>=1.0.2
colorama>=0.4.6
rich>=13.0.0
aiohttp>=3.9.0